src/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-agents>=1.2.0b4",
    "azure-ai-contentsafety>=1.0.0",
    "chainlit>=2.8.0",
    "openai>=1.108.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "semantic-kernel[azure]>=1.36.0",
]
